CoherentNoise.Common.AbstractSampler — Type

Supertype for all N-dimensional samplers.

CoherentNoise.Common.gen_image — Method

gen_image(sampler::AbstractSampler; <kwargs>)

Construct a 2-dimensional array of Images.RGB values, suitable for writing to disk as an image file.

Arguments

sampler::AbstractSampler: Any instance of a sampler. The sampler is sampled using each pixel coordinates as the X and Y input coordinates, and random Z and W coordinates for 3 and 4-dimensional samplers.

w::Integer=1024: The width in pixels of the image array to generate.

h::Integer=1024: The height in pixels of the image array to generate.

xbounds::NTuple{2,Float64}=(-1.0, 1.0): The bounds along the X axis to sample coordinates from. This remaps pixel coordinates to this range to be used for the input coordinates to sample with.

ybounds::NTuple{2,Float64}=(-1.0, 1.0): The bounds along the Y axis to sample coordinates from. This remaps pixel coordinates to this range to be used for the input coordinates to sample with.

colorscheme=nothing: A ColorSchemes.ColorScheme object to colorize the image with, or nothing.

CoherentNoise.Common.sample — Function

sample(sampler::AbstractSampler, x)
sample(sampler::AbstractSampler, x, y)
sample(sampler::AbstractSampler, x, y, z)
sample(sampler::AbstractSampler, x, y, z, w)

Sample from sampler with the supplied coordinates. The number of coordinates should match the dimensionality of the sampler type.

CoherentNoise.Patterns.Checkered — Type

A 2-dimensional sampler that generates a checkerboard pattern.

CoherentNoise.Patterns.Checkered — Method

Checkered(; seed=nothing)

Construct a sampler that outputs values in a checkerboard-like pattern when it is sampled from.

That is, output values will only ever be -1.0 or 1.0.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect the reproducibility of any samplers further down the pipeline.

CoherentNoise.Patterns.Constant — Type

A 1-dimensional sampler that emits a constant value.

CoherentNoise.Patterns.Constant — Method

Constant(; seed=nothing, value=0.0)

Construct a sampler that constantly outputs value each time it is sampled from.

This is useful for debugging and applications where you need to combine a constant value.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect the reproducibility of any samplers further down the pipeline.

value::Real=0.0: A constant value to emit each time this sampler is sampled from.

CoherentNoise.Patterns.Cylinders — Type

A 2-dimensional sampler that generates a pattern of concentric cylinders.

CoherentNoise.Patterns.Cylinders — Method

Cylinders(; seed=nothing, frequency=1.0)

Construct a sampler that outputs values that form a pattern representing concentric cylinders when it is sampled from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect the reproducibility of any samplers further down the pipeline.

frequency::Real=1.0: The frequency of the signal, which controls how small or large the cylinders are.

CoherentNoise.Patterns.PatternSampler — Type

Supertype for all N-dimensional pattern samplers.

CoherentNoise.Patterns.Spheres — Type

A 3-dimensional sampler that generates a pattern of concentric spheres.

CoherentNoise.Patterns.Spheres — Method

Spheres(; seed=nothing, frequency=1.0)

Construct a sampler that outputs values that form a pattern representing concentric spheres when it is sampled from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect the reproducibility of any samplers further down the pipeline.

frequency::Real=1.0: The frequency of the signal, which controls how small or large the spheres are.

CoherentNoise.Noise.NoiseSampler — Type

Supertype for all N-dimensional noise samplers.

CoherentNoise.Noise.ValueNoise.Value — Type

An N-dimensional sampler that generates value noise.

N must be an integer in the range [2, 4].

CoherentNoise.Noise.ValueNoise.Value — Method

Value{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional value noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.PerlinNoise.Perlin — Type

An N-dimensional sampler that generates Perlin "Improved" noise.

N must be an integer in the range [2, 4].

CoherentNoise.Noise.PerlinNoise.Perlin — Method

Perlin{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional Perlin "Improved" noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.SimplexNoise.Simplex — Type

An N-dimensional sampler that generates Simplex noise.

N must be an integer in the range [2, 4].

Note: ≥ 3 dimensions is patent-protected for certain applications. Consider using one of the OpenSimplex implementations, instead.

CoherentNoise.Noise.SimplexNoise.Simplex — Method

Simplex{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional Simplex noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.OpenSimplexNoise.OpenSimplex — Type

An N-dimensional sampler that generates OpenSimplex (legacy) noise.

N must be an integer in the range [2, 4].

Note: This version of OpenSimplex is considered "legacy" by its original author. Consider using one of the newer algorithms, OpenSimplex2 or OpenSimplex2S.

CoherentNoise.Noise.OpenSimplexNoise.OpenSimplex — Method

OpenSimplex{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional OpenSimplex noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.OpenSimplex2Noise.ImproveX — Type

Re-orient the noise space with the Y axis pointing down the main diagonal.

CoherentNoise.Noise.OpenSimplex2Noise.ImproveXY — Type

Re-orient the noise space to have better visual isotropy in the XY plane.

CoherentNoise.Noise.OpenSimplex2Noise.ImproveXYZ — Type

Re-orient a 4-dimensional noise space to be better suited for time-varied animations, where the W axis is time.

CoherentNoise.Noise.OpenSimplex2Noise.ImproveXZ — Type

Re-orient the noise space to have better visual isotropy in the XZ plane.

CoherentNoise.Noise.OpenSimplex2Noise.OpenSimplex2 — Type

An N-dimensional sampler that generates OpenSimplex2 noise, with an orientation of O.

OpenSimplex2 is not as smooth as OpenSimplex2S, although more performant.

N must be an integer in the range [2, 4].

O must be an Orientation type.

CoherentNoise.Noise.OpenSimplex2Noise.OpenSimplex2 — Method

OpenSimplex2{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional OpenSimplex2 noise when it is sampler from.

The noise space will be re-oriented using Standard orientation.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.OpenSimplex2Noise.OpenSimplex2 — Method

OpenSimplex2{N,O}(; seed=nothing)

Construct a sampler that outputs N-dimensional OpenSimplex2 noise when it is sampler from.

The noise space is re-oriented according by the supplied Orientation type O.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.OpenSimplex2Noise.Orientation — Type

Supertype of all orientations that can be applied to the OpenSimplex2 or OpenSimplex2S sampler types to re-orient the noise space.

CoherentNoise.Noise.OpenSimplex2Noise.Standard — Type

The standard orientation for OpenSimple2 and OpenSimplex2S.

CoherentNoise.Noise.OpenSimplex2SNoise.OpenSimplex2S — Type

An N-dimensional sampler that generates OpenSimplex2S noise, with an orientation of O.

OpenSimplex2S is smoother than OpenSimplex2, at the expense of being less performant.

N must be an integer in the range [2, 4].

O must be an Orientation type.

CoherentNoise.Noise.OpenSimplex2SNoise.OpenSimplex2S — Method

OpenSimplex2S{N}(; seed=nothing)

Construct a sampler that outputs N-dimensional OpenSimplex2S noise when it is sampler from.

The noise space will be re-oriented using Standard orientation.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.OpenSimplex2SNoise.OpenSimplex2S — Method

OpenSimplex2S{N,O}(; seed=nothing)

Construct a sampler that outputs N-dimensional OpenSimplex2S noise when it is sampler from.

The noise space is re-oriented according by the supplied Orientation type O.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

CoherentNoise.Noise.CellularNoise.CellFunction — Type

Supertype of all cellular noise cell functions.

CoherentNoise.Noise.CellularNoise.Cellular — Type

An N-dimensional sampler that generates cellular (Voronoi/Worley) noise.

N must be an integer in the range [2, 4]. M must be a DistanceMetric type. F must be a CellFunction type.

CoherentNoise.Noise.CellularNoise.Cellular — Method

Cellular{N}(; seed=nothing)
Cellular{N,M}(; seed=nothing)
Cellular{N,M,F}(; seed=nothing)

Construct a sampler that outputs N-dimensional cellular noise when it is sampler from.

M if supplied, must be a DistanceMetric type, which denotes the method to use for determining adjacent cells.

F if supplied, must be a CellFunction type, which denotes the "cell function" used in the cell distance calculation.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

jitter::Real=1.0: A Real number between 0.0 and 1.0, with values closer to one randomly distributing cells away from their grid alignment.

CoherentNoise.Noise.CellularNoise.Chebyshev — Type

Use the Chebyshev distance metric for determining the distance between two cells.

CoherentNoise.Noise.CellularNoise.DistanceMetric — Type

Supertype of all cellular noise distance metrics.

CoherentNoise.Noise.CellularNoise.Euclidean — Type

Use the Euclidean distance metric for determining the distance between two cells.

CoherentNoise.Noise.CellularNoise.Euclidean² — Type

Use the squared Euclidean distance metric for determining the distance between two cells.

CoherentNoise.Noise.CellularNoise.Manhattan — Type

Use the Manhattan distance metric for determining the distance between two cells.

CoherentNoise.Noise.CellularNoise.Minkowski4 — Type

Use the Minkowski4 (p=4) distance metric for determining the distance between two cells.

CoherentNoise.Fractals.Billow — Type

An N-dimensional sampler that generates a billowy fractal.

N must be an integer in the range [2, 4].

S may be an instance of any AbstractSampler, or an un-parameterized concrete type of one.

O denotes the number of octaves to apply for this fractal.

CoherentNoise.Fractals.Billow — Method

Billow{N}(; <kwargs>)

Construct a sampler that outputs an N-dimensional billowy fractal noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

source::Union{Type{<:AbstractSampler},AbstractSampler}=OpenSimplex2S: A sampler type or sampler instance to use as the source of the fractal. If this is an instance, its dimensionality must match N.

octaves=4: An integer between 1 and 32, denoting the number of octaves to apply.

frequency=1.0: The frequency of the first octave's signal.

lacunarity=2.0: A multiplier that determines how quickly the frequency increases for successive octaves.

persistence=0.5: A multiplier that determines how quickly the amplitude diminishes for successive octaves.

args...: Additional arguments to pass to the source if it is a type instead of an instance.

CoherentNoise.Fractals.FBM — Type

An N-dimensional sampler that generates a fractional Brownian motion fractal.

N must be an integer in the range [2, 4].

S may be an instance of any AbstractSampler, or an un-parameterized concrete type of one.

O denotes the number of octaves to apply for this fractal.

CoherentNoise.Fractals.FBM — Method

FBM{N}(; <kwargs>)

Construct a sampler that outputs an N-dimensional fractional Brownian motion fractal noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

source::Union{Type{<:AbstractSampler},AbstractSampler}=OpenSimplex2: A sampler type or sampler instance to use as the source of the fractal. If this is an instance, its dimensionality must match N.

octaves=4: An integer between 1 and 32, denoting the number of octaves to apply.

frequency=1.0: The frequency of the first octave's signal.

lacunarity=2.0: A multiplier that determines how quickly the frequency increases for successive octaves.

persistence=0.5: A multiplier that determines how quickly the amplitude diminishes for successive octaves.

args...: Additional arguments to pass to the source if it is a type instead of an instance.

CoherentNoise.Fractals.FractalSampler — Type

Supertype for all N-dimensional fractal samplers.

CoherentNoise.Fractals.Hybrid — Type

An N-dimensional sampler that generates a hybrid multifractal.

N must be an integer in the range [2, 4].

S may be an instance of any AbstractSampler, or an un-parameterized concrete type of one.

O denotes the number of octaves to apply for this fractal.

CoherentNoise.Fractals.Hybrid — Method

Hybrid{N}(; <kwargs>)

Construct a sampler that outputs an N-dimensional hybrid multifractal noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

source::Union{Type{<:AbstractSampler},AbstractSampler}=OpenSimplex2S: A sampler type or sampler instance to use as the source of the fractal. If this is an instance, its dimensionality must match N.

octaves=4: An integer between 1 and 32, denoting the number of octaves to apply.

frequency=1.0: The frequency of the first octave's signal.

lacunarity=2.0: A multiplier that determines how quickly the frequency increases for successive octaves.

persistence=0.25: A multiplier that determines how quickly the amplitude diminishes for successive octaves.

args...: Additional arguments to pass to the source if it is a type instead of an instance.

CoherentNoise.Fractals.Multifractal — Type

An N-dimensional sampler that generates a multifractal.

N must be an integer in the range [2, 4].

S may be an instance of any AbstractSampler, or an un-parameterized concrete type of one.

O denotes the number of octaves to apply for this fractal.

CoherentNoise.Fractals.Multifractal — Method

Multifractal{N}(; <kwargs>)

Construct a sampler that outputs an N-dimensional multifractal noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

source::Union{Type{<:AbstractSampler},AbstractSampler}=OpenSimplex2S: A sampler type or sampler instance to use as the source of the fractal. If this is an instance, its dimensionality must match N.

octaves=4: An integer between 1 and 32, denoting the number of octaves to apply.

frequency=1.0: The frequency of the first octave's signal.

lacunarity=2.0: A multiplier that determines how quickly the frequency increases for successive octaves.

persistence=0.5: A multiplier that determines how quickly the amplitude diminishes for successive octaves.

args...: Additional arguments to pass to the source if it is a type instead of an instance.

CoherentNoise.Fractals.Ridged — Type

An N-dimensional sampler that generates a ridged multifractal.

N must be an integer in the range [2, 4].

S may be an instance of any AbstractSampler, or an un-parameterized concrete type of one.

O denotes the number of octaves to apply for this fractal.

CoherentNoise.Fractals.Ridged — Method

Ridged{N}(; <kwargs>)

Construct a sampler that outputs an N-dimensional ridged multifractal noise when it is sampler from.

Arguments

seed::Union{UInt,Nothing}=nothing: A UInt used to seed the random number generator for this sampler, or nothing. If a seed is not supplied, one will be generated automatically which will negatively affect reproducibility.

source::Union{Type{<:AbstractSampler},AbstractSampler}=OpenSimplex2S: A sampler type or sampler instance to use as the source of the fractal. If this is an instance, its dimensionality must match N.

octaves=4: An integer between 1 and 32, denoting the number of octaves to apply.

frequency=1.0: The frequency of the first octave's signal.

lacunarity=2.0: A multiplier that determines how quickly the frequency increases for successive octaves.

persistence=1.0: A multiplier that determines how quickly the amplitude diminishes for successive octaves.

attenuation=2.0: The attenuation to apply to the weight of each octave.

args...: Additional arguments to pass to the source if it is a type instead of an instance.

CoherentNoise.Modifiers.ModifierSampler — Type

Supertype for all N-dimensional modifier samplers.

Base.:* — Method

*(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the product of the outputs samplers x and y.

Base.:* — Method

*(x::AbstractSampler, y::Real)

Construct a modifier sampler that outputs the product of the output of sampler x by the scalar y.

Base.:+ — Method

+(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the sum of the outputs of samplers x and y.

Base.:+ — Method

+(x::AbstractSampler, y::Real)

Construct a modifier sampler that outputs the sum of the output of sampler x and the scalar y.

Base.:- — Method

-(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the difference of the outputs of samplers x and y.

Base.:- — Method

-(x::AbstractSampler, y::Real)

Construct a modifier sampler that outputs the difference of the output of sampler x and the scalar y.

Base.:- — Method

-(x::AbstractSampler)

Construct a modifier sampler that outputs the negated output of sampler x.

Base.:/ — Method

/(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that performs division of the output of sampler x by the output of sampler y.

Base.:/ — Method

/(x::AbstractSampler, y::Real)

Construct a modifier sampler that performs division of the output of sampler x by the scalar y.

Base.:^ — Method

^(x::AbstractSampler, y::Real)

Construct a modifier sampler that raises the output of sampler x to the power of the scalar y.

Base.Math.clamp — Method

clamp(x::AbstractSampler, lo::AbstractSampler, hi::AbstractSampler)

Construct a modifier sampler that clamps the output of sampler x to be within the range of of output values from samplers lo and hi.

Base.Math.clamp — Method

clamp(x::AbstractSampler, lo::Real=-1.0, hi::Real=1.0)

Construct a modifier sampler that clamps the output of sampler x to be within the range of of the scalars lo and hi.

Base.abs — Method

abs(x::AbstractSampler)

Construct a modifier sampler that outputs the absolute value of its source when it is sampled from.

Base.copysign — Method

copysign(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the value of sampling from x with the sign copied from the value of sampling from y.

Base.max — Method

max(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the maximum value of the outputs of sampling from samplers x and y.

Base.min — Method

min(x::AbstractSampler, y::AbstractSampler)

Construct a modifier sampler that outputs the minimum value of the outputs of sampling from samplers x and y.

Base.muladd — Method

muladd(x::AbstractSampler, strength::Real=1.0, bias::Real=0.0)

Construct a modifier sampler that performs multiplies the output of sampler x by the scalar strength, followed by adding the scalar bias. sampler y.

CoherentNoise.Modifiers.cache — Method

cache(x::AbstractSampler)

Construct a modifier sampler that caches the set of the input coordinates and their corresponding output value of its source sampler. If the input coordinates differs from the previously cached output, the cache is invalidated and the new output is cached.

Caching is useful if a sampler is used as a source for multiple modifiers. Without caching, the duplicated input sources would redundantly compute the same outputs, which would be expensive, especially if long pipelines share a long subgraph.

CoherentNoise.Modifiers.curve — Method

curve(x::AbstractSampler, points::Vector{Pair{Float64,Float64}})

Construct a modifier sampler that outputs the result of sampling from x after remapping its output to an arbitrary curve.

The curve is defined by a Vector of Pairs given by points. Each pair of points represents an input and output number. The curve is a cubic spline, and so points must contain a list of four point pairs at a minimum. Additionally, no two point pairs can contain the same input point value.

When sampling from sampler x, the output is evaluated using the curve data, and maps it to a new output value.

CoherentNoise.Modifiers.mix — Method

mix(x::AbstractSampler, y::AbstractSampler, t::AbstractSampler)

Construct a modifier sampler that outputs the result of linearly interpolating the output of samplers x and y by the output of sampler t.

CoherentNoise.Modifiers.mix — Method

mix(x::AbstractSampler, y::AbstractSampler, t::Scalar)

Construct a modifier sampler that outputs the result of linearly interpolating the output of samplers x and y by the scalar t.

CoherentNoise.Modifiers.rotate — Method

rotate(source::AbstractSampler; x::Real=0.0, y::Real=0.0, z::Real=0.0)

Construct a modifier sampler that rotates the input coordinates of sampler source around the origin before sampling from it.

The coordinate system is assumed to be left-handed.

The angle of rotation is specified in radians for the corresponding axis given by x, y, and z.

CoherentNoise.Modifiers.scale — Method

scale(source::AbstractSampler, scale::Real=1.0)

Construct a modifier sampler that uniformly scales the input coordinates of sampler source by the scalar scale before sampling from it.

CoherentNoise.Modifiers.scale — Method

scale(source::AbstractSampler; x::Real=1.0, y::Real=1.0, z::Real=1.0, w::Real=1.0)

Construct a modifier sampler that scales the input coordinates of sampler source before sampling from it.

Each axis can be scaled independently with x, y, z, or w.

CoherentNoise.Modifiers.select — Method

select(x::AbstractSampler, y::AbstractSampler; <kwargs>)

Construct a modifier sampler that outputs either the out of sampler x or y, depending on the output of sampler z.

If the output of sampler z is within the range denoted by min and max, the output of sampler y is chosen. If the output of sampler z is outside of this range, the output of sampler x is chosen.

Arguments

min::Real=-1.0: A real number between -1.0 and 1.0 defining the lower bound of the selection range.

max::Real=1.0: A real number between -1.0 and 1.0 defining the upper bound of the selection range.

falloff::Real=1.0: A real number between 0.0 and 1.0 specifying the smoothness of the transition.

CoherentNoise.Modifiers.terrace — Method

terrace(x::AbstractSampler, points::Vector{Pair{Float64,Float64}}; invert::Bool=false)

Construct a modifier sampler that outputs the result of sampling from x after remapping its output to a terrace-forming curve.

The curve is defined by a Vector of Float64s given by points. Each point represents an input and output number.

When sampling from sampler x, the output is evaluated using the curve data, and maps it to a new output value.

Arguments

invert::Bool=false: Specify whether the curve is inverted between control points.

CoherentNoise.Modifiers.translate — Method

translate(source::AbstractSampler; x::Real=0.0, y::Real=0.0, z::Real=0.0, w::Real=0.0)

Construct a modifier sampler that translates the input coordinates of sampler source along a specified vector, with a direction and magnitude given by the coordinates x, y, z, and w.

CoherentNoise.Modifiers.turbulence — Method

turbulence(s1::AbstractSampler; s2::AbstractSampler; <kwargs>)

Construct a modifier sampler that displaces the input coordinates of sampler s1 by the output of sampler s2 with a fractional Brownian motion fractal applied to it.

Sampler s2's input coordinates are randomly generated using the seed of sampler s1.

Arguments

frequency::Real=1.0: The frequency of the fractal signal to apply to sampler s2.

roughness::Real=3: The number of octaves of the fractal to apply to sampler s2.

power::Real=1.0: A scaling factor that is applied to the displaced result before sampling from sampler s1.

CoherentNoise.Modifiers.warp — Method

warp(source::AbstractSampler; <kwargs>)

Construct a modifier sampler that performs domain warping of the sampler source before sampling from it.

Domain warping feeds the output of other samplers to the input of a sampler. For this modifier, each input coordinate can specify a different sampler to warp with.

If a sampler is not supplied for x, y, z, or w, a sampler that outputs a constant zero will be used instead.

Arguments

x::AbstractSampler=Constant(): A sampler to warp the X axis by.

y::AbstractSampler=Constant(): A sampler to warp the Y axis by.

z::AbstractSampler=Constant(): A sampler to warp the Z axis by.

w::AbstractSampler=Constant(): A sampler to warp the W axis by.